# Add MultiVisitor to reduce the need to re-walk the ast