Found this wile working on #94

This shows ~30% speedup:

```
tjackson@tjackson-ThinkPad-X1-Carbon-5th:~/workspace/golang/src/github.com/jacksontj/promxy/test$ benchcmp before after 
benchmark                                                               old ns/op     new ns/op     delta
BenchmarkEvaluations/benchdata/aggregators.test/promxy_remoteread-4     3985050       2839665       -28.74%

benchmark                                                               old allocs     new allocs     delta
BenchmarkEvaluations/benchdata/aggregators.test/promxy_remoteread-4     11245          11256          +0.10%

benchmark                                                               old bytes     new bytes     delta
```